examples: Add 'cookbook' example
This is based on the wiki page
https://developer.mozilla.org/en-US/docs/Mozilla/Projects/SpiderMonkey/JSAPI_Cookbook
but it is less likely to bitrot if it can be compiled. The idea of this
file is not to do anything in particular but just to contain snippets of
code in small functions that can be browsed educationally or copied into
your program.