build: Update clang-format config file, and reformat examples
This new config is copied from the root of the mozilla-central codebase
(but with the NS_* macros removed, which are irrelevant here.) All the
code examples are reformatted, using
`clang-format -i -style=file examples/*.cpp`.

In addition, all the exceptions marked with 'clang-format off' are
removed. (This was mostly used for JSClassOps anyway, which will later
get replaced with JS::DefaultGlobalClassOps.)

Closes #15.