examples: Add boilerplate code file to use in examples
This encapsulates the minimal context and global initialization code
used by most examples. It should be kept and straight-forward as
possible as it is expected to be read and understood by embedders.

The definition of DefaultGlobalClassOps matches that of SM 65+. It uses
standard resolve hooks for the builtin standard types.