Add generated bytecodes documentation
This copies the Python scripts from the Gecko source tree that generate
the bytecodes MDN page at
https://developer.mozilla.org/en-US/docs/Mozilla/Projects/SpiderMonkey/Internals/Bytecode
and modifies them to produce valid Markdown.

Adds a preamble copied from
https://developer.mozilla.org/en-US/docs/Mozilla/Projects/SpiderMonkey/Internals/Bytecodes

Also adds the generated result.

This document will need to be regenerated once per ESR release. It
requires a checkout or unpacked tarball of the SpiderMonkey source.